+++ doc/release-notes.md
@@ -20,6 +20,9 @@ Updated RPCs
   `window_final_block_height`.
 - The `getnetworkinfo` and `getpeerinfo` commands now contain
   a new `servicesnames` field with decoded network service flags.
+- `walletprocesspsbt` and `walletcreatefundedpsbt` now include BIP 32
+  derivation paths by default for public keys if we know them. This can be
+  disabled by setting `bip32derivs` to `false`.
 
 Updated settings
 ----------------